Working Perspective matrix and scene

+++ Rin.Shading/DebugInfo.cs
@@ -25,16 +25,16 @@ public struct DebugInfo(string file, uint beginLine, uint beginCol, uint endLine
 
     public static DebugInfo operator +(DebugInfo left, DebugInfo right)
     {
-        var minBeginLine = float.Min(left.BeginLine, right.BeginLine);
+        var minBeginLine = uint.Min(left.BeginLine, right.BeginLine);
         var minBeginCol = left.BeginLine == minBeginLine && right.BeginLine == minBeginLine
-            ? float.Min(left.BeginCol, right.BeginCol)
+            ? uint.Min(left.BeginCol, right.BeginCol)
             : left.BeginLine == minBeginLine
                 ? left.BeginCol
                 : right.BeginCol;
 
-        var maxEndLine = float.Max(left.EndLine, right.EndLine);
+        var maxEndLine = uint.Max(left.EndLine, right.EndLine);
         var maxEndCol = left.EndLine == maxEndLine && right.EndLine == maxEndLine
-            ? float.Max(left.EndCol, right.EndCol)
+            ? uint.Max(left.EndCol, right.EndCol)
             : left.EndLine == maxEndLine
                 ? left.EndCol
                 : right.EndCol;
